+++ resolwe/flow/tests/test_backend.py
@@ -30,7 +30,7 @@ class BackendTest(TestCase):
 
     def tearDown(self):
         for data in Data.objects.all():
-            data_dir = os.path.join(settings.FLOW_EXECUTOR['DATA_PATH'], str(data.id))
+            data_dir = os.path.join(settings.FLOW_EXECUTOR['DATA_DIR'], str(data.id))
             shutil.rmtree(data_dir, ignore_errors=True)
 
     def test_manager(self):